% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtm-plot-helpers.R
\name{.draw_hexagon}
\alias{.draw_hexagon}
\title{Draw a Filled Hexagon}
\usage{
.draw_hexagon(cx, cy, r, col, border = NA)
}
\arguments{
\item{cx, cy}{Center coordinates.}

\item{r}{Radius (circumradius).}

\item{col}{Fill color.}

\item{border}{Border color.}
}
\description{
Draw a Filled Hexagon
}
\keyword{internal}
