% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtm-internal.R
\name{.gtm_bi}
\alias{.gtm_bi}
\title{Beta Initialization from Nearest-Neighbor Distances}
\usage{
.gtm_bi(Y)
}
\arguments{
\item{Y}{Numeric matrix (K x D), mapped points.}
}
\value{
Numeric scalar, initial beta.
}
\description{
Beta Initialization from Nearest-Neighbor Distances
}
\keyword{internal}
