% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtm-internal.R
\name{.gtm_lbf}
\alias{.gtm_lbf}
\title{Linear Basis Functions}
\usage{
.gtm_lbf(X)
}
\arguments{
\item{X}{Numeric matrix (K x L).}
}
\value{
Numeric matrix (K x (L+1)).
}
\description{
Returns X with a bias column appended.
}
\keyword{internal}
