% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtm-internal.R
\name{.gtm_pts}
\alias{.gtm_pts}
\title{Generate 1D Latent Grid Points}
\usage{
.gtm_pts(M)
}
\arguments{
\item{M}{Integer, number of grid points.}
}
\value{
Numeric matrix (M x 1).
}
\description{
Creates M evenly spaced points in [-1, 1].
}
\keyword{internal}
