% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtm-internal.R
\name{.gtm_resp3}
\alias{.gtm_resp3}
\title{Wrapper for .gtm_resp in Mode 0}
\usage{
.gtm_resp3(DIST, beta, D)
}
\arguments{
\item{DIST}{Numeric matrix (K x N).}

\item{beta}{Numeric > 0.}

\item{D}{Integer.}
}
\value{
List with llh and R.
}
\description{
Wrapper for .gtm_resp in Mode 0
}
\keyword{internal}
