% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtm-internal.R
\name{.gtm_ri}
\alias{.gtm_ri}
\title{Random Weight Initialization}
\usage{
.gtm_ri(TT, FI)
}
\arguments{
\item{TT}{Numeric matrix (N x D), training data.}

\item{FI}{Numeric matrix (K x (M+1)), basis function activations.}
}
\value{
Numeric matrix ((M+1) x D).
}
\description{
Initializes weight matrix W randomly, scaled to match data variance.
}
\keyword{internal}
