% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtm-internal.R
\name{.gtm_sort}
\alias{.gtm_sort}
\title{Sort Responsibilities}
\usage{
.gtm_sort(R)
}
\arguments{
\item{R}{Numeric matrix (K x N), responsibilities.}
}
\value{
Sorted responsibility matrix.
}
\description{
Sort Responsibilities
}
\keyword{internal}
