\name{gtm.ppd1}
\alias{gtm.ppd1}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Latent space posterior probability distribution for a given data point }
\description{
This function calculates the posterior probability 
distribution induced in the latent space of a trained GTM model 
for a given data point, and returns it in a format suitable for  
2D graphic plotting routines.
}
\usage{
gtm.ppd1(tt, Y, beta, X)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{tt}{a point in the data space; 1-by-D }
  \item{Y}{ centres of the Gaussian mixture generated by the GTM in the data space, Y = FI*W; K-by-D }
  \item{beta}{ variance of Gaussian mixture; scalar  }
  \item{X}{ latent sample  }
}
\value{
 List with components
  \item{X}{latent sample}
  \item{P}{posterior distribution over latent space given data point t}
}
\seealso{  \code{\link{gtm.pmn}}, \code{\link{gtm.pmd}} }
\examples{
  lv <- gtm.demo()
  gtm.ppd1(lv$TT[1,], lv$Y, lv$beta, lv$X)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{cluster}
\keyword{models}
