\name{gtm.ppd2}
\alias{gtm.ppd2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Latent space posterior probability distribution for a given data point }
\description{
This function calculates the posterior probability distribution induced in the latent space of a trained GTM model for a given data point, and returns it in a format suitable for  3D graphic plotting routines
}
\usage{
gtm.ppd2(tt, Y, beta, X, xDim, yDim)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{tt}{a point in the data space; 1-by-D }
  \item{Y}{ centres of the Gaussian mixture generated by the GTM in the data space, Y = FI*W; K-by-D }
  \item{beta}{ variance of Gaussian mixture; scalar  }
  \item{X}{ latent sample  }
  \item{xDim,yDim}{ number of points along the 2 dimensions of the latent space meshgrid sample }
}
\value{
 List with components
  \item{X,Y }{latent sample; if the latent space is 2D, \code{X} and \code{Y} 
      are mesh matrices}
  \item{Z }{posterior distribution over latent space given data point t}
}
\seealso{   \code{\link{gtm.pmn}}, \code{\link{gtm.pmd}}}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{cluster}
\keyword{models}% ..ONLY ONE.. keyword per line
