% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtm-methods.R
\name{print.gtm}
\alias{print.gtm}
\title{Print Method for GTM}
\usage{
\method{print}{gtm}(x, ...)
}
\arguments{
\item{x}{A \code{"gtm"} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for GTM
}
