% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtm-methods.R
\name{summary.gtm}
\alias{summary.gtm}
\title{Summary Method for GTM}
\usage{
\method{summary}{gtm}(object, ...)
}
\arguments{
\item{object}{A \code{"gtm"} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns a list with summary statistics.
}
\description{
Summary Method for GTM
}
